GUFusion/requirements.txt
asgiref==3.2.7
astroid==2.4.1
click==7.1.2
colorama==0.4.3
coverage==5.1
dj-static==0.0.6
Django==3.0.6
django-adminlte2==0.4.1
django-jet==1.0.8
django-stdimage==5.1.1
gunicorn==20.0.4
isort==4.3.21
joblib==0.14.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
model-mommy==2.0.0
nltk==3.5
Pillow==7.1.2
progressbar2==3.51.3
psycopg2-binary==2.8.5
pylint==2.5.2
python-utils==2.4.0
pytz==2020.1
regex==2020.5.7
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
textblob==0.15.3
toml==0.10.0
tqdm==4.46.0
wrapt==1.12.1
